+++ config/errors/__init__.py
@@ -1,3 +1,6 @@
+from pathlib import Path
+
+
 class ConfigurationError(Exception):
     """An exception risen for invalid configuration."""
 
@@ -7,6 +10,6 @@ class ConfigurationOverrideError(ConfigurationError):
 
 
 class MissingConfigurationFileError(ConfigurationError, FileNotFoundError):
-    def __init__(self, file_path: str) -> None:
+    def __init__(self, file_path: Path) -> None:
         super().__init__(f"Missing configuration file: {file_path}")
         self.missing_file_path = file_path